+++ type-query/src/main/kotlin/dev/sayaya/handbook/interfaces/database/R2dbcTypeEntity.kt
@@ -29,7 +29,7 @@ data class R2dbcTypeEntity(
         type.primitive(primitive)
         type.parent(parent)
         type.attributes(attributes.toTypedArray())
-        rev?.let { type.rev(it) }
+        rev?.let { type.rev(it.toDouble()) }
         return type
     }
 }
